fix issue #135: constant param update and use node id in rq worker

+++ PYTHON/WATCH/watch.py
@@ -85,35 +85,32 @@ def run(**kwargs):
                 try:
                     job = Job.fetch(r_obj['ALL_JOBS'][i], connection=r)
                 except Exception:
-                    print(traceback.format_exc())
-                job.delete()
-
+                    print('Error in cancelling existing job: ', Exception, traceback.format_exc())
+                if job is not None:
+                    job.delete()
 
     def report_failure(job, connection, type, value, traceback):
         print(job, connection, type, value, traceback)
-
-    def jid(n):
-        return get_redis_obj(jobset_id)['ALL_JOBS'][n]
-
+    
     try:
         for n in topological_sorting:
             cmd = nodes_by_id[n]['cmd']
             ctrls = nodes_by_id[n]['ctrls']
             func = getattr(globals()[cmd], cmd)
-            job_id = 'JOB_' + cmd + '_' + uuid.uuid1().__str__()
+            job_id = nodes_by_id[n]['id']  # 'JOB_' + cmd + '_' + uuid.uuid1().__str__()
             s = ' '.join([STATUS_CODES['JOB_IN_RQ'], cmd.upper()])
             r_obj = get_redis_obj(jobset_id)
             prev_jobs = r_obj['ALL_JOBS'] if 'ALL_JOBS' in r_obj else {};
             r.set(jobset_id, dump({
                 **r_obj,'SYSTEM_STATUS': s, 'ALL_JOBS': {
-                    **prev_jobs, cmd: job_id
+                    **prev_jobs, job_id: job_id
                 }
             }))
             send_to_socket({
                 'SYSTEM_STATUS': s,
                 'jobsetId': jobset_id
             })
-            r.rpush(jobset_id+'_ALL_NODES', cmd.upper())
+            r.rpush(jobset_id+'_ALL_NODES', job_id)
             if len(list(DG.predecessors(n))) == 0:
                 q.enqueue(func,
                         # TODO: have to understand why the SINE node is failing for few times then succeeds
@@ -126,8 +123,7 @@ def run(**kwargs):
             else:
                 previous_job_ids = []
                 for p in DG.predecessors(n):
-                    prev_cmd = DG.nodes[p]['cmd']
-                    prev_job_id = jid(prev_cmd)
+                    prev_job_id = DG.nodes[p]['id']
                     previous_job_ids.append(prev_job_id)
                 q.enqueue(func,
                         retry=Retry(max=100),
